Remove storage box ID and spaces from label
The pnp4nagios Performance Data specification is given in "Nagios
plugins" format. This is the expected format:

  'label'=value[UOM];[warn];[crit];[min];[max]

Although the notes state that 'the single quotes for the label are
optional. Required if spaces, = or ' are in the label', putting quotes
around the "Storage Box #{}" label did not work for me.

For details, see: https://docs.pnp4nagios.org/pnp-0.6/perfdata_format

And as the storage box ID is given in the description of the check, I
decided to remove the ID and the spaces from the label, instead I just
use "storagebox_usage".

Closes: viafintech/check_hetzner_storage_box#7